backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
anthropic==0.8.1
python-dotenv==1.0.0
eventlet==0.33.3  # For Socket.IO
psycopg2-binary==2.9.9  # PostgreSQL driver
gunicorn==21.2.0  # For production deployment
python-engineio==4.8.0  # Explicitly set version for compatibility
python-socketio==5.10.0  # Explicitly set version for compatibility 